feat: refactor storage structure to use repositories for better organization and maintainability
